+++ R/estimation.R
@@ -4,8 +4,8 @@
 #'
 #' @param object an object of class `pqrBayes'.
 #' @param coefficient the vector of quantile regression coefficients under a linear model or the matrix of true varying coefficients evaluated on the grid points under a varying coefficient model.
-#' @param u.grid the vector of grid points under a varying coefficient model. When fitting a linear (quantile) regression model, u.grid = NULL.
-#' @param model the model to be fitted. Users can choose "linear" for a linear model or "VC" for a varying coefficient model.
+#' @param u.grid the vector of grid points under a varying coefficient model. When fitting a linear (quantile) regression model or group LASSO, u.grid = NULL.
+#' @param model the model to be fitted. Users can choose "linear" for a linear model, "VC" for a varying coefficient model or "group for group LASSO.
 #' @usage estimation.pqrBayes(object,coefficient,u.grid=NULL,model="linear")
 #' @return  an object of class `pqrBayes.est' is returned, which is a list with components:
 #' \item{error}{mean square error or integrated mean square errors and total integrated mean square error.}
@@ -19,7 +19,7 @@
 #' y=data$y
 #' e=data$e
 #' coeff = data$coeff
-#' fit1=pqrBayes(g,y,u=NULL,e,quant=0.5,spline=NULL,model="linear")
+#' fit1=pqrBayes(g,y,u=NULL,e,d = NULL,quant=0.5,spline=NULL,model="linear")
 #' estimation=estimation.pqrBayes(fit1,coeff,model="linear")
 
 #'
@@ -31,8 +31,11 @@ estimation.pqrBayes = function(object,coefficient,u.grid=NULL,model="linear"){
   else if(model=="linear"){
     pqrBayes.est = estimation_lin(object,coefficient)
   }
+  else if(model=="group"){
+    pqrBayes.est = estimation_lin(object,coefficient)
+  }
   else{
-    stop("model should be either VC or linear")
+    stop("model should be either VC, linear or group")
   }
   class(pqrBayes.est) = "pqrBayes.est"
   return(pqrBayes.est)