+++ R/nonrobust_g.R
@@ -1,10 +1,10 @@
-NonRobust_lin <- function(g, y, e,iterations,sparse,debugging){
-
+NonRobust_g <- function(g, y, e, d,iterations,sparse,debugging){
+  
   p = dim(g)[2]
   
   x = cbind(1,g)
   n = length(y); 
-
+  
   if(!is.null(e)){
     q = dim(e)[2]
     hatAlpha=rep(1,(1+q))
@@ -17,8 +17,8 @@ NonRobust_lin <- function(g, y, e,iterations,sparse,debugging){
     w = matrix(1,n,1)
   }
   hatBeta = rep(1,p)
-  hatInvTauSq = rep(1,p)
-  hatLambdaSq=1
+  hatInvTauSqStar = rep(1,p/d)
+  hatLambdaSqStar=1
   hatSigmaSq=1
   aStar=1 
   bStar=1
@@ -31,8 +31,10 @@ NonRobust_lin <- function(g, y, e,iterations,sparse,debugging){
   
   progress = ifelse(debugging, 10^(floor(log10(iterations))-1), 0)
   if(sparse){
-    fit=BLSS(g, y, w, iterations, hatAlpha, hatBeta, hatInvTauSq, invSigAlpha0, hatPi, hatLambdaSq, hatSigmaSq, aStar, bStar, alpha, gamma, sh1, sh0, progress)}
-  else{fit=BL(g, y, w, iterations, hatBeta, hatAlpha, hatInvTauSq, invSigAlpha0, hatLambdaSq, hatSigmaSq, aStar, bStar, alpha, gamma, progress)}
+    fit=BGLPointMass(g, y, w, p/d, d, iterations, hatAlpha, hatBeta, hatInvTauSqStar,invSigAlpha0, hatPi,
+                     hatLambdaSqStar, hatSigmaSq, aStar, bStar, alpha, gamma, sh1, sh0, progress)
+  }
+  else{fit=BGL(g, y, w, p/d, d, iterations, hatBeta, hatAlpha, hatInvTauSqStar, invSigAlpha0, hatLambdaSqStar, hatSigmaSq, aStar, bStar, alpha, gamma, progress)}
   out=list(fit=fit,iterations=iterations)
   return(out)
 }