+++ R/pqrBayes.R
@@ -1,23 +1,29 @@
-#' fit Bayesian penalized quantile regression for linear and varying coefficient models
+#' fit Bayesian penalized quantile regression for linear, group LASSO, or varying coefficient models
 #' 
 #' @keywords models
 #' @param g the matrix of predictors (subject to selection). Users do not need to specify an intercept which will be automatically included. 
 #' @param y the response variable. The current version only supports the continuous response.
-#' @param u a vector of effect modifying variable of the quantile varying coefficient model. When fitting a linear model, u = NULL.
+#' @param u a vector of effect modifying variable of the quantile varying coefficient model. When fitting a linear model or group LASSO, u = NULL.
 #' @param e a matrix of clinical covariates not subject to selection.
 #' @param quant the quantile level specified by users. The default value is 0.5.
+#' @param d a positive integer denotes the group size. When fitting a linear model or varying coefficient model, d = NULL.
 #' @param iterations the number of MCMC iterations. The default value is 10,000.
 #' @param burn.in the number of burn-in iterations. If NULL, the first half of MCMC iterations will be used as burn-ins.
-#' @param spline a list of the number of interior knots (kn) for B-spline and the degree of B-spline basis (degree). When fitting a linear model, spline = NULL.
+#' @param spline a list of the number of interior knots (kn) for B-spline and the degree of B-spline basis (degree). When fitting a linear model or group LASSO, spline = NULL.
 #' @param robust logical flag. If TRUE, robust methods will be used. Otherwise, non-robust methods will be used. The default value is TRUE.
 #' @param sparse logical flag. If TRUE, spike-and-slab priors will be adopted to impose exact sparsity on regression coefficients. Otherwise, Laplacian shrinkage will be adopted. The default value is TRUE.
-#' @param model the model to be fitted. Users can specify "linear" for a linear model and "VC" for a varying coefficient model.
+#' @param model the model to be fitted. Users can specify "linear" for a linear model, "group" for group LASSO and "VC" for a varying coefficient model.
 #' @param hyper a named list of hyper-parameters. The default value is NULL.
 #' @param debugging logical flag. If TRUE, progress will be output to the console and extra information will be returned. The default value is FALSE.
 #'
-#' @details The quantile regression model described in "\code{\link{data}}" is:
+#' @details 
+#' The quantile regression model described in "\code{\link{data}}" is:
 #' \deqn{Y_{i}=\sum_{k=1}^{q} E_{ik} \beta_k +\sum_{j=0}^{p}X_{ij}\gamma_j +\epsilon_{i},}
 #' where \eqn{\beta_k}'s are the regression coefficients for clinical covariates and \eqn{\gamma_j}'s are the regression coefficients of \eqn{\boldsymbol X}.
+#' 
+#' The group LASSO model described in "\code{\link{data}}" is:
+#' \deqn{Y_{i}=\sum_{k=1}^{q} E_{ik} \beta_k + X_{i0}\gamma_0+ \sum_{j=1}^{m}\boldsymbol{X_{ij}^\top}\boldsymbol{\gamma_j} +\epsilon_{i},}
+#' where \eqn{\beta_k}'s are the regression coefficients for clinical covariates and \eqn{\boldsymbol{\gamma_j} = (\gamma_{j1},\dots,\gamma_{jd})^\top} is the vector of regression coefficients of the \eqn{n \times d} matrix \eqn{\boldsymbol X_{j}}.
 #' 
 #' The quantile varying coefficient model described in "\code{\link{data}}" is:
 #' \deqn{Y_{i}=\sum_{k=1}^{q} E_{ik} \beta_k +\sum_{j=0}^{p}\gamma_j(V_i)X_{ij} +\epsilon_{i},}
@@ -45,20 +51,39 @@
 #' y=data$y
 #' e=data$e
 #' 
-#' fit1=pqrBayes(g,y,u=NULL,e,quant=0.5,spline=NULL,model="linear")
+#' fit1=pqrBayes(g,y,u=NULL,e,d=NULL,quant=0.5,model="linear")
 
 #' \donttest{
 #'
 #' ## non-sparse
 #' sparse=FALSE
-#' fit2=pqrBayes(g,y,u=NULL,e,quant=0.5,spline=NULL,sparse = sparse,model="linear")
+#' fit2=pqrBayes(g,y,u=NULL,e,d=NULL,quant=0.5,spline=NULL,sparse = sparse,model="linear")
 #' 
 #' ## non-robust
 #' robust = FALSE
-#' fit3=pqrBayes(g,y,u=NULL,e,quant=0.5,spline=NULL,robust = robust,model="linear")
+#' fit3=pqrBayes(g,y,u=NULL,e,d=NULL,quant=0.5,spline=NULL,robust = robust,model="linear")
 #' 
 #' }
 #' 
+#' ## The group LASSO model
+#' data(data)
+#' data = data$data_group
+#' g=data$g
+#' y=data$y
+#' e=data$e
+#' 
+#' fit1=pqrBayes(g,y,u=NULL,e,d=3,quant=0.5,model="group")
+#' \donttest{
+#'
+#' ## non-sparse
+#' sparse=FALSE
+#' fit2=pqrBayes(g,y,u=NULL,e,d=3,quant=0.5,spline=NULL,sparse = sparse,model="group")
+#' 
+#' ## non-robust
+#' robust = FALSE
+#' fit3=pqrBayes(g,y,u=NULL,e,d=3,quant=0.5,spline=NULL,robust = robust,model="group")
+#' 
+#' }
 #' ## The quantile varying coefficient model
 #' data(data)
 #' data = data$data_varying
@@ -68,7 +93,7 @@
 #' e=data$e
 #'
 #' spline = list(kn=2,degree=2)
-#' fit1=pqrBayes(g,y,u,e,quant=0.5,spline = spline, model="VC")
+#' fit1=pqrBayes(g,y,u,e,quant=0.5,spline = spline,model="VC")
 #'
 #' \donttest{
 #'
@@ -96,16 +121,17 @@
 #' {\emph{Statistics in Medicine}, 39: 617– 638} \doi{10.1002/sim.8434}
 
 
-pqrBayes <- function(g, y, u=NULL, e,quant=0.5, iterations=10000, burn.in=NULL,spline, robust=TRUE,sparse=TRUE, model="linear",hyper=NULL,debugging=FALSE){
+pqrBayes <- function(g, y, u=NULL, e, d=NULL, quant=0.5, iterations=10000, burn.in=NULL,spline=NULL, robust=TRUE,sparse=TRUE, model="linear",hyper=NULL,debugging=FALSE){
   
   if(model=="VC"){
     kn = as.integer(spline$kn); degree = as.integer(spline$degree)
     fit = pqrBayes_vc(g, y, u, e,quant, iterations, burn.in, kn, degree,robust,sparse, hyper,debugging)
   }else if(model=="linear"){
-    fit = pqrBayes_lin(g, y, e,quant, iterations, burn.in, robust,sparse, hyper,debugging)
-  }
-  else{
-    stop("model should be either VC or linear")
+    fit = pqrBayes_lin(g, y, e, quant, iterations, burn.in, robust,sparse ,hyper,debugging)
+  }else if(model=="group"){
+    fit = pqrBayes_g(g, y, e, d, quant, iterations, burn.in, robust,sparse ,hyper,debugging)
+  }else{
+    stop("model should be either VC, linear or group.")
   }
   return(fit)
 }